[component-syntax] Clean up hints for component params
Summary:
Component parameters are expected to have annotations, so there is no point in trying to compute hints for them based on context. This diff cleans up the `~hints` argument for the two relevant methods in Name_def.

It also adds `root` variant for an unannotated catch parameters (`UnannotatedParameter`).

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D46537637

fbshipit-source-id: 94a229757c6e67d12aa776e996cf683998b23a58