[refactor] Use named fields for Type.Property
Summary:
The loc part of property was first added in D6213491 for get-def and find-ref purposes. However, over the years, the invariants are not well maintained, and the fields are not always populated, which leads to get-def jump to weird places.

This diff doesn't attempt to fix the issue. Instead, it converts the tuple type to named struct type, to make it much easier to audit, and much less likely to be passed a wrong location in the future.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D46543211

fbshipit-source-id: 3693b71d25cc60407925d839941d86ffcc6f2ccc