[get-def] Propagate key_loc in slice_utils to make property get-def reliably jump to name locations
Summary:
With the refactors in the previous diff, it's clear that we are not propagating the key_loc to mapped properties in slice_utils. We should do it and significantly improve the accuracy and consistency of get-def requests involving these object utility types.

Changelog: [ide] For property go-to-definition for objects types produced by utility types, we will now reliably jump to the original object definition's field name locations

Reviewed By: samwgoldman

Differential Revision: D46544945

fbshipit-source-id: 2d7ccdec4f20a6ef5b4ce322e93cc0863379c7ab