Populate key_loc for object spread when the later field completely overrides the earlier field
Summary:
In cases like `{...A, ...B}`, when `B` has non-optional prop `foo` that completely overrides the prop `foo` in `A`, we should use the key_loc of `foo` `B` as the key_loc of `foo` in the spreaded object type.

This allows us to always point to prop name locations in get-def in situations like this, which is likely the most common object spreads when overrides happens.

For optional props when there is no complete overrides, we should keep track of both locations, but we don't have the structure for this yet. I will followup in a later diff.

Changelog: [ide] For spreaded object types when a later non-optional field completely overrides an easier ones, go-to-definition will always jump to the field name location of the later field's type definition.

e.g.

```
type A = {foo: number};
type B = {foo: string};
//         ^ jump here
declare var obj: {...A,...B};

obj.foo;
//   ^ get-def
```

Reviewed By: samwgoldman

Differential Revision: D46604352

fbshipit-source-id: b8b4c447742439c47f314f8aa15849beab5d19b6