adapt hints to type guards
Summary:
When we decompose hints to extract function parameters, we create "template" functions with placeholder tvars. These are compared against concrete input functions to instantiate the placeholders.

When the input function type has a type guard, it is important that the template *also* is a type guard and of the *same kind* (ie. refining the same parameter), otherwise we risk causing errors during hint decomposition, which can throw off the entire process.

To ensure we don't walk into this kind of errors, we match the structure of the target input function type by:

1/ matching the names of parameters. In most contexts `(mixed) => boolean` and `(x: mixed) => boolean` are equivalent type. But if these types are type guards, `(x: mixed) => x is string` and `(mixed) => x is string` are not the same: the latter type is invalid.

2/ matching the presence of the predicate. Even though we might only care about a hint on the parameter type, if we don't take type predicate into account, then the constraint between the template for the function and the input type would be
```
(x: tvar) => any ~> (x: mixed) => x is number
```
which is also an error. So, instead, we need to produce
```
(x: tvar) => x is any ~> (x: mixed) => x is number
```
This is what this diff does. (1) is largely taken care of in the previous diff. This diff implements (2).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D46505041

fbshipit-source-id: a37134b33973b2735fec559c2f9fe5e098473608