type guard consistency checks
Summary:
When defining a predicate function of the form
```
function foo(x: mixed): x is T { body; }
```
we make sure that two things hold:

1. the type of `x` after executing `body` and applying all refinements is a subtype of `T`
2. `x` has not been reassigned by another value. This is important because we need to be able to connect that variable that is passed as argument to `foo` with the result of calling `foo`. To guarantee this, we check for two things:
2a. `x` is not assigned directly in `foo`.
2b. `x` is not havoced in `foo`.

Examples of inconsistent type guard functions:
```
function bad_1(x: mixed): x is 1 { // error 2 ~> 1
  return x === 2;
}

function bad_2a(x: mixed): 1 {
  x = 1;
  return x === 1; // error 'x' is reassinged in the previous line
}

function bad_2b(x: mixed): 1 {
  function inner() {
    x = 1;
  }
  inner();
  return x === 1; // error 'x' is havoced in 'inner'
}
```

To implement this we extend the information we gather during the name_resolver pass. Respectively, for each one of the above goals:
1. We use `synthetic_read` on `x` (the type predicate parameter) right after the return expression (or expression body) to get the refined version of `x`. This has the form of a (list of) `write_loc`, which is really expected to be a series of `Refinement`s wrapping an initial `Write` from the function's parameter.
2. To make sure we accurately connect the result of the function with the parameter:
2a. If the structure `Refinement (... Refinement (... Write param_loc)..)` from (1) does not hold, we error with a message that blames the locations that write to `x` within `foo`.
2b. We extend name_resolver to keep track of the call that causes a `havoc_env` so that when we record that in the case of a type guard function, we can more accurately blame the call that causes the havoc.

Note that all of these errors are actually raised in statement.ml `mk_fun_sig`, because that's where we raise other predicate related errors, and also it is easier to connect the error messages with information about the parameter that is expected to be refined. Since these errors involve typing of the function body, it is important that these happen in the `reconstruct_ast` part of `mk_fun_sig`, after the function body has been checked. The rest of the predicate checks were moved there in D46449292 in preparation of this diff.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D46449290

fbshipit-source-id: e5c4c338faec24479b010fab99b8833631a796b2