Simplify constructor special-case for method unbinding autofix
Summary:
For the method-unbinding error, we have an autofix to change the method to a
class property. However, we don't want to propose changing the constructor.

We don't support string literal properties, so we can remove this case and just
look at identifiers. Furthermore, the check for string literal properties was
wrong: it looked at `raw` which includes quotes and will therefore never be
equal to the string `constructor` (without quotes).

Reviewed By: mroch

Differential Revision: D46573725

fbshipit-source-id: 3df3e4efaf121111b25a0082a539929c91578da4