Fully resolve EvalT result when its inputs are resolved outside of implicit instantiation
Summary:
This diff starts to fully resolve the results of EvalT. The core idea is that, if all of the inputs do not contain unresolved tvars, then nothing happening in the future could potentially change the result of this EvalT. We should seal the type so that it doesn't get widened.

However, the implementation of this diff is less principled, in order to minimize the error diff. I also excluded implicit instantiations for no good reasons, except that it helps to reduce the error diff. I only did it for conditional types and mapped types, which are currently unreleased.

I choose these two types first because I want them to not inherit any bad behavior from the start, and let individual people who want to adopt these better types to deal with potential behavior differences with the old $Call and $ObjMap, instead of shipping a broken version and fix it later with all other util types. I plan to gradually roll it out to all destructors over time.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D44757036

fbshipit-source-id: 5e5771abb4b8dd91b62ad674448be27bce2493fa