[component-syntax] Improve reason for component config.
Summary:
We were re-using the reason created for the component as the reason for the computed config. This diff adds a new reason desc that we can use to wrap the RComponentReason (and eventually an RComponentTypeReason) that specifies the *props* of that component. I'm intentionally using the props terminology instead of config because config is a flow-only term and is not helpful for our users.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D46545543

fbshipit-source-id: b57e916c6e99e8c71c9d87b1f189b3fa1890d132