Extract type of instance from config in components
Summary:
This diff is, perhaps, getting ahead of myself, but to prepare for the switch from FunT to AbstractComponentT for components, this diff extracts the instance type from the "ref" property of component configs. It still raises an error on "ref"s, but it also extracts the type, and flows it into `React$Ref<fresh tvar>` in order to extract the instance (this is on the theory that, as I understand it, a component like `component Foo(ref: Ref<Bar>)` will have an instance type of `Bar`). We don't do anything with this instance type yet, but when we switch to AbstractComponents we'll be able to drop it right in.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D46601342

fbshipit-source-id: f61895c19509736a6b1cf1cdde4bb48bd2f49370