encode sentinel refinements for type guards
Summary:
It is possible to emulate sentinel checks using type guards, by including in the type guard position an (inexact) object type with just the key that is intended to be used as sentinel. For example, in
```
type A = { tag: 'A', foo: number };
type B = { tag: 'B', foo: string };

declare var x: A | B;

declare function sentinel(x: mixed): x is {tag: 'A', ...};

if (sentinel(x)) {
  (x: A); // okay A ~> A
  (x: B); // error A ~> B
}
```
the guard only includes a "tag" key with the value that corresponds to the `A` case.

To make this work with the existing filtering mechanism, we need to make sure that when `B` is filtered against `{ tag: 'A', ... }`, we return `empty`, instead of `B & { tag: 'A', ...}`, which would be the fallback case before.

With the setup we had until now this was not the case. Consider the steps that the algorithm follows from D46606214:
1. runtime tag non-overlap check: both types are tagged as `ObjTag`, so there is potential overlap
2. quick subtyping fails in both directions
3. (normal) subtyping in both directions

So we would end up with the fallback (4), which is the intersection creation.

The only check that could realistically change to accommodate this filtering is the tag non-overlap check. In this diff, we encode the presence of a sentinel key/value as payload `sentinel` to the `ObjTag`. `sentinel` maps sentinel strings to a value. We only consider values that are string, numeric or boolean literals. This is similar to how `SentinelProp` check works.

We also update the comparator operator over `TypeTag.t`s to consider two `ObjTag`s different iff they differ in at least one of their common sentinel keys. So:
* `type A = { tag: 'A', foo: number }` and `{tag: 'A', ...}` overlap, but
* `type B = { tag: 'B', foo: string }` and `{tag: 'A', ...}` do not overlap.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D46634549

fbshipit-source-id: bdfe2def930ced14fb75f0604d0db1b7d5eee99d