better intersection of input and type guard
Summary:
Before this diff, when an input type `T` was refined against a type guard `S`, the result was typically `T & S`. While this is a correct result it often leads to spurious results. For example, consider
```
declare function isBoolean(x: mixed): x is boolean;
const x = 1;
if (isBoolean(x)) {
  (x: empty);
}
```
while we know that `number & boolean` should be equivalent to `empty`, the above cast would fail with a speculation error of the form
```
Cannot cast `x` to empty because: [incompatible-cast]
 * Either number [1] is incompatible with empty [2].
 * Or boolean [3] is incompatible with empty [2].
```
Notably, this is not equivalent with the behavior you'd get with the equivalent code
```
const x = 1;
if (typeof x === "boolean") {
  (x: empty);
}
```
which does not have any errors, because the incoming `number` type is refined to `empty` when it is tag-checked with "boolean".

We need, therefore, a way to improve the way we intersect types as the result of type guard refinement and bring it closer to the level of the current predicate based refinement.

This diff tries to achieve this in two ways:

A. We introduce the notion of *runtime tags* to describe sets of values that have no overlap, and use this to return `empty` when the input and type guard sets of values have no overlap.

As far as implementation is concerned, a new data type `TypeTag.t` is introduced in Type_filtering, as well as a `tag_of_t` function that conservatively associates a `Type.t` with a set of `TypeTag.t`s. Note that at the time that we call `tag_of_t` most complex types (e.g. `EvalT`, `OpenT`, `TypeAppT`) have boiled away thanks to concretization:
* On the input type side (so for type `T` from earlier) this happens because constraints of the form `_ ~> PredicateT _` appear pretty late in the Flow_js pattern match.
* On the type guard side (so for type `S`) this happens explicitly by calling `possible_concrete_types_for_inspection` immediately when calling the `intersect` function.

B. We call full-blown subtyping checking (via `SpeculationKit.try_singleton_throw_on_failure`) right after `quick_subtype`. If `S <: T` we return `S`, if `T <: S`, we return `T`. Otherwise, we return `S & T`.

So, overall, the filtering algorithm consists of
1. runtime tag non-overlap check
2. quick subtyping
3. (normal) subtyping
4. (fallback) intersection creation

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D46606214

fbshipit-source-id: 911f8fe2b4933ae5866064f4bb6d456e9f3b0a80