Unbreak autofix related to missing import
Summary:
We got user report that the auto-import is broken since the last release. Upon closer inspection, auto-import via auto-complete still works, but auto-import via an existing error is broken.

It turns out that it's caused by a crash here for ALoc.t -> Loc.t transformation:

https://github.com/facebook/flow/blob/1db40803eb2b8448264a8fa02ec8f57cd2d7ff90/src/services/code_action/autofix_class_member_access.ml#L26-L32

The bad code was there when the support for this auto fix is first added, but it's probably only broken recently, likely due to some of my typing related changes for imports.

To trigger the issue, you need to be in a file with existing import and an missing global error. Unsurprisingly, this is not covered by our existing test suite, so the error is never caught.

This diff fixes the issue by doing a proper loc conversion through loc_of_aloc function from a reader. I also added a regression test to ensure it actually works.

I will do a patch release for this.

Changelog: [ide] Fix a crash when we are trying to produce a quick fix for missing imports. The autofix for missing imports broken in the previous release would be fixed.

Reviewed By: mroch

Differential Revision: D46677044

fbshipit-source-id: c81bf072f40a7a02b9d49806a1f9bce181932e09