type guard refinement negation
Summary:
This diff adds support for the negation of a type-guard predicate.

Problem statement:

Given input type t1 and type guard t2, let T1 and T2 the sets of values represented by t1 and t2, respectively. We are looking for: T1 ∩ T2', where T2' is the complement of T2.

Also, since unions in t1 are broken up before the refinement, T1 can be expressed as T1_1 U ... U T1_n, and so the outgoing set becomes:

(T1_1 ∩ T2') U ... U (T1_n ∩ T2')

We handle this case by checking if t1_i <: t2, and if that's the case we return `empty`, otherwise we return the input type t1_i.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D46651382

fbshipit-source-id: ae4fb123a724c674795d8191d604b70951c65b3d