Better description for array spread reversal related error message
Summary:
When the source of the tuple is coming from a function during array rest reversal, the reason description will still say function type, which is not very helpful and even confusing. This diff changes the reason desc to RTupleType or RArray based on the reversal result.

Changelog: [errors] Error messages related to array rest in generic functions are slightly improved

Reviewed By: samwgoldman

Differential Revision: D46702014

fbshipit-source-id: dbe6dd80e23b02e9839bcd6de3052af8eb652e1a