Introduce Type.normalized_property
Summary:
In a lot of places, we create Type.property object with dummy key_loc. This worries me since it makes it hard to audit whether a `key_loc = None` is truly unnecessary because downstream uses don't inspect it all, or it might be some get-def accuracy issue.

This diff creates a new type specifically for the purpose of passing it to places that inspect properties that don't need the key loc. These uses are usually type checking code that only requires polarity and type information, like GetPropKit or perform_lookup_action. With this change, the number of `key_loc = None` for properties have significantly decreased. (Most of the remaining ones are legit, e.g. we truly don't have a key_loc for the synthetic default constructor)

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D46612813

fbshipit-source-id: 10dae5cf831de9d8cf9b15cb4434bfba9d0f325d