Correct error on unannotated component params
Summary: This fixes an error where an unannotated component param would raise an internal error. The issue here was that the `UnannotatedParam` name def would not trigger the resolution of the `FunctionParamLoc` name_def on the same location, which a regular param (annotated or contextual) would. This diff fixes, and adds a missing annotation error in its place.

Reviewed By: jbrown215

Differential Revision: D46774618

fbshipit-source-id: b6cb7d4641eb85927e2100a5a72d38aba2780125