[component-syntax] Fix optional property calculation for component params
Summary:
We were only marking a component param to be optional if it had a default, but params can also be optional if they're declared to be optional. This diff checks to see if the optional flag is set when computing the rest/param types for the config of a component and sets them to be optional if necessary.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D46780349

fbshipit-source-id: c395366940d0b71e51ae9593e289c3b0a6e6a3e8