[component-syntax] Fix component declaration param parsing
Summary:
Previously, the parsing for component params did not consider optional params without a rename and allowed nameless props:

```
component OptionalNoRename(x?: number) {...}
component NamelessDestructure({bar, baz}) {...}
```

Now we allow the first example and error on the second. I was able to consolidate several cases to make this happen, and the location ranges for the parameters now take into account the annotations.

To fix this, I changed the ast representation of the parameter name to no longer be an option, which requires changes to various modules in the codebase. This matches the parsing behavior for components in hermes-parser, which requires a param name. I then had to address the parser cases that introduced the `None` option, and I got rid of it by consolidating several cases into one simplified case, which also fixes the issue with non-renamed optional parameters.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D46772942

fbshipit-source-id: b4b1324e0f379829925fdf726874094a7d9e8b1f