[component-syntax] Rest param constraints
Summary:
The rest param for components should be non-optional and an object type. A simple way to check this at the declaration is to flow its type to `{...}` and use a customized error message to explain that the rest param must be an object type and non-optional.

I prefer this over using object_kit, which is built to transform most types with an object-like representation into an object, as it would introduce a lot of special cases.

I'm not sure if the way I'm going about handling the special use-op is good form or not.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D46653819

fbshipit-source-id: d9b8d15a1af74930b8629bc375877172bc67f9c0