Error when creating a JSX intrinsic element, if there is a component-like value in scope with the same (lowercase) name
Summary:
This diff is intended to address the issue that a React component may be assigned to a lower-case variable, after which a user may try to create an element from that variable without realizing that lower-case JSX titles are always intrinsics. One way to address this is to prevent such assignments from happening, but this might be more complicated than it sounds.

This diff addresses the issue from the other direction: when instantiating an intrinsic, we see if there are any values in scope with the same name as the intrinsic. If there are, and those values have a type that could be instantiated as an element, we raise a lint error.

The requirement that the variable be something that could be a component is because there are lots of ways that this name overlap could happen innocently: it would be annoying to, for example, error on a case like this:

```
const div = <div>Hi</div>;
return <div>{div}</div>
```
However, if the `div` variable here was a React component, we would error.

This design is enabled by the fact that in the env_builder, we don't special case lower-case JSX titles--we still create an environment read for them that reads from the variable, so we can just use the existing machinery here to look up that entry and see what it is. This was an oversight in the design of the env builder, so if we decide *not* to use this diff, we should fix that oversight.

Also note that the error message here needs improvement!

Reviewed By: jbrown215

Differential Revision: D46498387

fbshipit-source-id: 41c688ae11291de2d13f31afcde88c0ec61eab2a