fix in type guards for methods
Summary:
In some final testing I noticed that type guards were not recorded properly for methods. The reason is most likely that `mk_methodtype` had `predicate` as an optional parameter which was not passed in by `Func_sig.methodtype`.

This diff fixes this by making this parameter a required optional predicate, and passing it appropriately when needed. It also adds tests with predicates in classes and objects.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D46820698

fbshipit-source-id: 36f6f657a95c2f9eff0caf834eb8da4758ba2e05