Extract type of instance from config in components, type_sig version
Summary:
This diff does the same work as D46601342, but in the type sig. This is a little more complicated here, because as far as I know (and please correct me if I'm wrong, samwgoldman or others who know this system better) there isn't a simple way in the type sig to perform the equivalent of the `T ~> React$Ref<S>` flow in order to get the type of `S`. Instead we add a simple type destructor to defer that flow until we're in flowjs.

For consistency, we now use this type destructor during typechecking as well.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D46653837

fbshipit-source-id: 04e013d9ee2cd70040480c4455ce36c5617b1356