When creating array type from React children, don't set tuple types of ArrayAT if there is a spread
Summary:
The second argument of `ArrayAT` is `Some [...]` if we know the specific elements of the array when it is viewed as a tuple.

In this code, we have some JSX children spread (the `Some spread`). We take this into account when constructing the first argument of `ArrayAT`, which is the `T` when the array is viewed as an `Array<T>`. However, it doesn't make sense that we should construct a view of this array type with known tuple elements, since we don't know what's in the spread.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D46960153

fbshipit-source-id: 72ff2dda30156bcb682355a4b202b94c0f5cf6db