fix opaque types in type-guard filtering
Summary:
Consider the underlying or super type of an opaque type when computing tags during type-guard filtering.

Note that we respect the opaque type transparency rule so filtering based on underlying type only happens in the same file where the opaque type is defined. This causes the following code to fail to be refined properly because we cannot determine if `void` is included in `OpaqueString`, since this type is defined in another file.
```
declare function isTruthy<A>(value: ?A): value is A;
declare var x: void | OpaqueOrString; // imported opaque type (see added test case)

if (isTruthy(x)) {
  (x: OpaqueOrString);
}
```
This works as expected if `OpaqueOrString` is defined locally.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision:
D47034752

------------------------------------------------------------------------
(from fd867e987952b948e577232c863520a9414d5337)

fbshipit-source-id: c8313589a7d75c804cc8edeaf1a4592e1462c42a