fix autocompleting keywords eating text
Summary:
the range used for keywords in autocomplete included the imaginary `AUTO332` sigil, which caused accepting a keyword suggestion to eat the next 7 characters of the line.

I noticed it with this example:

```
class C e| {
}
```
accepting `extends` erases the space and `{` resulting in

```
class C extends|
}
```

also, since it was using the same range for `insert` and `replace`, this also didn't work:

```
class C |implements D {
}
```

triggering autocomplete here, you should be able to shift-enter on `extends` to replace `implements`. instead, it always inserted, resulting in `extendsimplements`.

Changelog: [fix] Fix a bug where autocompleting a keyword deleted the next 7 characters on the same line

Reviewed By: SamChou19815

Differential Revision: D47115336

fbshipit-source-id: f9031df2ff763b3a0f8995969fe5168434a3a271