don't rank autoimports by usage when ranking by usage is off
Summary:
`Fuzzy_path.search` currently uses a weighted fuzzy score. Weights are only meant to be used when ranked imports are enabled (the weights are how often the import is used).

This diff adds a flag so that only the fuzzy score is used by default.

Changelog: [fix] "Fix" sorting of autoimports. An experimental option to sort autoimports by how often they are imported was partially influencing the sorting even when the option was not enabled. This was likely an improvement, but it was a bug to not obey the option.

Reviewed By: SamChou19815

Differential Revision: D47116551

fbshipit-source-id: 8bc01c61bfbf9ce71e2ca154f50949fa76f016e6