Fully resolve EvalT resposition result if the EvalT result will be fully resolved
Summary:
We can resolve in these cases, similar to we also resolve reposition result for OpenT when the OpenT is at least resolved.

I found this potential improvement when I'm looking at why some inputs to conditional types are not resolved.

Changelog: [internal] There are some slight error diffs, but the impact is so small and it depends so much on implementation detail that it doesn't worth mentioning it

Reviewed By: jbrown215

Differential Revision: D47179395

fbshipit-source-id: 837b4f6c44a91e93913ca0156cedfe9dc5346093