Fix flow-parser size issue
Summary:
The ppx library is being unnecessary included in the runtime code. Because we fork sedlex, it's also being included, but somehow that's able to get correctly DCEed. Anyways, this diff fixes it by not including it.

Fixes https://github.com/facebook/flow/issues/9044

Changelog: [parser] We fixed the parser size regression since 0.210.0.

Reviewed By: mvitousek

Differential Revision: D47200237

fbshipit-source-id: 51c7a64ae2672e266025ff15a7156844475577c4