set status background to red when disconnected
Summary:
users have reported that they have trouble realizing when the Flow server is stopped. the vscode status bar item says "server is stopped", but it's not all that obvious.

this diff turns it red (if the extension supports `window/showStatus` with a `backgroundColor` param).

Changelog: [internal] the flowtype.flow-for-vscode extension doesn't support this yet

Reviewed By: gkz

Differential Revision: D47251823

fbshipit-source-id: 31fecfdb58c5b10330e014cab4ca44a8ada7e0b7