show lesser-used exact autoimports starting in 4th place
Summary:
in D47251875, i remove the exact-match bonus and order the results by score and then weight. this causes exact matches and prefix matches to have the same scores, so sorted by weight.

in practice, this can cause the exact matches to end up at the bottom of the list. for example, internally we have hundreds of things that start with `f` that are used thousands of times, and one export named `f` that is used once.

```
1) foo (100)
2) foobar (99)
3) floof (98)
4) far (97)
...
99) flop (2)
100) f (1)
```

if `f` is what you wanted, it's in 100th place and you'll probably never find it. if you can't find it, and can't type anything else to filter the results more (because it's already an exact match), then you're out of luck.

so this diff puts exact matches in 4th place. you might get something like (usage counts in parens)

```
1) foo (100)
2) foobar (99)
3) floof (98)
4) f (1)
5) far (97)
...
99) flop (2)
```

based on our telemetry, <1% of accepted completions are the 4th place result, so this is not displacing any more relevant results. people are likely skimming the list and typing more until the right item is first. in that case, they can easily skim over that entry and type one more character to filter it away.

Changelog: [internal] ranked imports are experimental

Reviewed By: SamChou19815

Differential Revision: D47251871

fbshipit-source-id: c0a48cdc0c4f46a267f9185deb21f80bb321b475