fix overweight fuzzy score for uppercase patterns
Summary:
Uppercase characters in a common prefix of a fuzzy match are given an undue boost.

This manifests when trying to sort without boosting full matches. When searching for `F` with `boost_full_match = false`, the score for `F` is still +2 the score for `Foo` even though they should be equal.

Ports https://github.com/microsoft/vscode/pull/187149 to our ocaml fork

Reviewed By: SamChou19815

Differential Revision: D47251873

fbshipit-source-id: 32e5102747cf3bb4c9ec613f613f546c52dbc596