tweak autoimport ranking
Summary:
separates the fuzzy score and weight (usage count) in autoimport ranking.

before, we computed a weighted score as `usage count + (fuzzy score ^ 4 * 1000) + kind bonus`. since `fuzzy score` is between 0 and 1, `fuzzy score ^ 4 * 1000` is between 0-1000, and so very badly-scoring things with more than 1000 usages can still dominate. this makes it hard to find things even with an exact match.

instead, we'll move to sorting by score, then by weight, but with special handling of exact matches that allows an inexact but highly-used result to outrank an exact but rarely-used result (this'll happen in followup diffs).

this diff on its own mostly regresses things, because the `Fuzzy_path` algorithm is pretty bad for this use case. in particular, it gives higher score to short results, so because we're sorting by score first, a very short, very rare result will beat out a longer very common result. followup diffs will fix this by replacing that algorithm with the `Fuzzy_score` algorithm we (and vscode) use elsewhere.

Changelog: [internal] only impacts ranked imports which are experimental

Reviewed By: SamChou19815

Differential Revision: D47251874

fbshipit-source-id: a9e6a34f105a6bc86c2b8a0fff4ea24c12bc4e62