Special case object merging for cjs exports to retain original def_loc
Summary:
Similar to the logic of using the definition loc of the original definition of ExportRef, we should do similar things for CommonJS exports.

In CommonJS, the equivalent is `module.exports = {foo}`. In this diff, I special cases the treatment for these object props with shorthand syntax, so that we can later use the information to directly jump to the definition location of foo.

This diff is the first diff to make use of the `preferred_def_locs` field. We populate it with the original def_loc of names exported via `module.exports = {...}`

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D46418581

fbshipit-source-id: 1c7d56e19996521a226b24a574c192dbfd97718f