[get-def] Always jump to the original definition name for required names
Summary:
With the improvement in the previous diff, we no longer need to exclude require pattern destructuring from the special behavior of using `Get_def_request.Member`.

With this change, we have a positive change that these require patterns will consistently jump to the name of the original definition.

Changelog: [ide] For imported names using require, go-to-definition will now jump to the location of the original definition's name.

e.g.
```
// a.js
const a = 3;
//    ^ jump here
module.exports = {a}

// b.js
const {a}=require('./a')
//     ^ request here
```

Reviewed By: gkz

Differential Revision: D46418578

fbshipit-source-id: f2bef8de9ad252347ebdaad0ec93e9bb6f1f56b8