prevent infinite loops in get-def
Summary:
`GetDef_js` has a `Depth` checker that's supposed to avoid runaway get-defs, but it doesn't seem to be working.

there are 3 conditions that should make it abort: pointing at itself, pointing outside the file, or getting stuck in a loop for 9999 iterations.

we have a repro where doing get-def on a parse error that gets recovered as an empty identifier, interacts somehow with all the other parse errors in the file that also are empty identifiers. somehow they avoid the depth limit.

I tried lowering the limit to 100 and it still didn't stop. But either way, spinning for 100 or 10,000 iterations is pointless. we should just detect the loop and abort right away.

so, this diff doesn't fix whatever is going on with these cyclical get-defs, but it does keep a `LocSet` and bails out as soon as it's seen a cycle.

Changelog: [fix] Prevent hover and go to definition requests that infinitely loop

Reviewed By: jbrown215

Differential Revision:
D47316595

------------------------------------------------------------------------
(from f2d8a476b3170096f4d6fc0d0aa2697feeb95727)

fbshipit-source-id: 00bfae1b7c1657c3822485fa0be3894ce6c24cb9