[tests] Make ide_unmerged_deps less flaky
Summary:
The test is asserting that get-def request that only depends on fast_dependent is not blocked by checking of slow_dependent, which has a sleep to intentionally slow it down.

This kind of test that depends on timeout is flaky regardless, but we can make it less flaky by asserting it finishes in 5 seconds instead of 2, while making the sleep 10s.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D47339031

fbshipit-source-id: e0fddab6cd3c33974bea306bd65aab1225e18bb7