[tests] Remove unnecessary `module.system=haste` in tests
Summary:
In quite a lot of test suites that have `module.system=haste`, the main purpose is to test single module type checking, and quite a lot of tests don't have import in them. It doesn't make sense to have them in config.

This is not a exhaustive cleanup. I just cleanup some of them until I got lazy.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D47455132

fbshipit-source-id: adbdf528b34e69d6f6559e8d9d9de45bff4594ae