Remove distributed typecheck intern project
Summary:
This was a really cool inten project, but didn't quite make it to production. We
don't have near-term plans to drive distributed type checking, so let's remove
this code, to make it easier to iterate on local checking without worrying about
the distributed case.

For now, this diff keeps the options flag around, but it's unused.

Reviewed By: gkz

Differential Revision: D47534504

fbshipit-source-id: 8db93b01f86676ddab619938476aa6b4a134f3ad