Remove check service abstraction around shared memory
Summary:
To support distributed checking, we created a little abstraction so that
type checking a single file did not depend directly on shared memory. Instead,
it depended on a "Reader" abstraction that could be implemented by shared memory
or some other storage -- for distcheck, that would have been CAS.

However, since we don't have near-term plans to support distributed checking,
we can simplify our lives by taking a strong dependency on shared memory. This
makes the existing use cases a bit easier to hack on, and we can add the
abstraction back if it's necessary later.

Reviewed By: gkz

Differential Revision: D47535294

fbshipit-source-id: 6d0a7aad47661647207277ff8dd2abc1b6dd1b13