Change resource file cjs_export to have no location
Summary:
Currently, the location of a resource file's export is the location of the
import "source." This means we create a distinct type for each require of a
given module.

I want to move toward a model where we construct single module type per
dependency, instead of per require. This will help simplify the code to resolve
requires to the dependency module's type. (See next diff in stack)

Note that we still reposition the type, so type errors are still helpful.
This does affect the "references" section of a type error. Since the def loc is
locationless, we provide no reference. Before, we would provide a useless
reference to the require source location.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D47615813

fbshipit-source-id: 1e0b2885249c5ee9b7c2fc25a9f7128baaf537c1