Make file sig consistent with statement pass on require with type args
Summary:
In statement.ml, we error on require calls with type args (`require<T>('foo')`),
and do not need the dependency module's type. The file sig analysis treats this
code as a dependency.

This diff makes the file sig analysis consistent, so it does not include the
dependency. The typeapp_call test output changes show the effect of this diff:
we do not error for missing modules for require<T>(...) calls.

In practice, the extra dependency is not a problem. This change makes the next
diff easier to review, since that diff has the same effect on the typeapp_call
test.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D47620157

fbshipit-source-id: c280f8c022ca21f71bf994dc2c0dd502cc949c9a